--- src/main/java/com/noplanb/character/controller/CharacterController.java
@@ -1,11 +1,0 @@
-package com.noplanb.character.controller;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-@Controller
-@RequiredArgsConstructor
-@RequestMapping("/api/v1")
-public class CharacterController {
-}
