--- src/main/java/com/noplanb/character/repository/CharacterRepository.java
@@ -1,9 +1,0 @@
-package com.noplanb.character.repository;
-
-import com.noplanb.character.domain.Character;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface CharacterRepository extends JpaRepository<Character,Long> {
-}
