--- src/main/java/com/noplanb/character/service/CharacterService.java
@@ -1,11 +1,0 @@
-package com.noplanb.character.service;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-@Service
-@RequiredArgsConstructor
-@Transactional(readOnly = true)
-public class CharacterService {
-}
