--- src/main/java/com/noplanb/common/BaseEntity.java
@@ -1,32 +1,0 @@
-package com.noplanb.common;
-
-import jakarta.persistence.*;
-import lombok.Getter;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-import java.time.LocalDateTime;
-
-@Getter
-@MappedSuperclass
-@EntityListeners(AuditingEntityListener.class)
-public abstract class BaseEntity {
-
-    @CreatedDate
-    @Column(updatable=false)
-    private LocalDateTime createdAt;
-
-    @LastModifiedDate
-    private LocalDateTime modifiedAt;
-
-    @Enumerated(value = EnumType.STRING)
-    private Status status = Status.ACTIVE;
-
-    public void updateStatus(Status status){
-        this.status = status;
-    }
-
-
-
-}
