--- src/main/java/com/noplanb/common/Status.java
@@ -1,6 +1,0 @@
-package com.noplanb.common;
-
-public enum Status {
-    ACTIVE,
-    DELETED
-}
