+++ src/main/java/com/noplanb/domain/character/repository/CharacterRepository.java
@@ -1,0 +1,9 @@
+package com.noplanb.domain.character.repository;
+
+import com.noplanb.domain.character.domain.Character;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface CharacterRepository extends JpaRepository<Character,Long> {
+}
