+++ src/main/java/com/noplanb/domain/common/Status.java
@@ -1,0 +1,6 @@
+package com.noplanb.domain.common;
+
+public enum Status {
+    ACTIVE,
+    DELETED
+}
