+++ src/main/java/com/noplanb/domain/item/controller/ItemController.java
@@ -1,0 +1,13 @@
+package com.noplanb.domain.item.controller;
+
+import io.swagger.v3.oas.annotations.tags.Tag;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+@Controller
+@RequiredArgsConstructor
+@RequestMapping("/api/v1")
+@Tag(name = "ItemController", description = "ItemControllerController입니다.")
+public class ItemController {
+}
