+++ src/main/java/com/noplanb/domain/item/domain/ItemStatus.java
@@ -1,0 +1,6 @@
+package com.noplanb.domain.item.domain;
+
+public enum ItemStatus {
+    equip,
+    unequip
+}
