+++ src/main/java/com/noplanb/domain/item/domain/ItemType.java
@@ -1,0 +1,8 @@
+package com.noplanb.domain.item.domain;
+
+public enum ItemType {
+    head,
+    face,
+    clothes,
+    background
+}
