+++ src/main/java/com/noplanb/domain/item/repository/ItemRepository.java
@@ -1,0 +1,9 @@
+package com.noplanb.domain.item.repository;
+
+import com.noplanb.domain.item.domain.Item;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface ItemRepository extends JpaRepository<Item,Long> {
+}
