+++ src/main/java/com/noplanb/domain/quest/controller/QuestController.java
@@ -1,0 +1,13 @@
+package com.noplanb.domain.quest.controller;
+
+import io.swagger.v3.oas.annotations.tags.Tag;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+@Controller
+@RequiredArgsConstructor
+@RequestMapping("/api/v1")
+@Tag(name = "QuestController", description = "QuestControllerController입니다.")
+public class QuestController {
+}
