+++ src/main/java/com/noplanb/domain/quest/domain/Quest.java
@@ -1,0 +1,29 @@
+package com.noplanb.domain.quest.domain;
+
+import com.noplanb.domain.character.domain.Character;
+import com.noplanb.domain.common.BaseEntity;
+import jakarta.persistence.*;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import static jakarta.persistence.GenerationType.IDENTITY;
+
+@Entity
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class Quest extends BaseEntity {
+    @Id
+    @GeneratedValue(strategy = IDENTITY)
+    private Long id;
+    private String contents;
+    private Long exp;
+    private Boolean isComplete;
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "character_id")
+    private Character character;
+
+    public void updateCharacter(Character character) {
+        this.character = character;
+    }
+}
