+++ src/main/java/com/noplanb/domain/user/controller/UserController.java
@@ -1,0 +1,13 @@
+package com.noplanb.domain.user.controller;
+
+import io.swagger.v3.oas.annotations.tags.Tag;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+@Controller
+@RequiredArgsConstructor
+@RequestMapping("/api/v1")
+@Tag(name = "UserController", description = "UserControllerController입니다.")
+public class UserController {
+}
