+++ src/main/java/com/noplanb/domain/user/repository/UserRepository.java
@@ -1,0 +1,9 @@
+package com.noplanb.domain.user.repository;
+
+import com.noplanb.domain.user.domain.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface UserRepository extends JpaRepository<User,Long> {
+}
