+++ src/main/java/com/noplanb/domain/user/service/UserService.java
@@ -1,0 +1,11 @@
+package com.noplanb.domain.user.service;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service
+@RequiredArgsConstructor
+@Transactional(readOnly = true)
+public class UserService {
+}
