+++ src/main/java/com/noplanb/global/config/SwaggerConfig.java
@@ -1,0 +1,24 @@
+package com.noplanb.global.config;
+
+import io.swagger.v3.oas.models.Components;
+import io.swagger.v3.oas.models.OpenAPI;
+import io.swagger.v3.oas.models.info.Info;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class SwaggerConfig {
+    @Bean
+    public OpenAPI openAPI() {
+        return new OpenAPI()
+                .components(new Components())
+                .info(apiInfo());
+    }
+
+    private Info apiInfo() {
+        return new Info()
+                .title("growMe API") // API의 제목
+                .description("나 키우기에 대한 api입니다.") // API에 대한 설명
+                .version("1.0.0"); // API의 버전
+    }
+}
