--- src/main/java/com/noplanb/item/domain/Item.java
@@ -1,32 +1,0 @@
-package com.noplanb.item.domain;
-
-import com.noplanb.character.domain.Character;
-import com.noplanb.common.BaseEntity;
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import static jakarta.persistence.GenerationType.IDENTITY;
-
-@Entity
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Item extends BaseEntity {
-    @Id
-    @GeneratedValue(strategy = IDENTITY)
-    private Long id;
-    private String item_name;
-    @Enumerated(EnumType.STRING)
-    private ItemType itemType;
-    @Enumerated(EnumType.STRING)
-    private ItemStatus itemStatus;
-    private Long requiredLevel;
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "character_id")
-    private Character character;
-
-    public void updateCharacter(Character character) {
-        this.character = character;
-    }
-}
