--- src/main/java/com/noplanb/item/domain/ItemStatus.java
@@ -1,6 +1,0 @@
-package com.noplanb.item.domain;
-
-public enum ItemStatus {
-    equip,
-    unequip
-}
