--- src/main/java/com/noplanb/item/domain/ItemType.java
@@ -1,8 +1,0 @@
-package com.noplanb.item.domain;
-
-public enum ItemType {
-    head,
-    face,
-    clothes,
-    background
-}
