--- src/main/java/com/noplanb/item/repository/ItemRepository.java
@@ -1,9 +1,0 @@
-package com.noplanb.item.repository;
-
-import com.noplanb.item.domain.Item;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface ItemRepository extends JpaRepository<Item,Long> {
-}
