--- src/main/java/com/noplanb/item/service/ItemService.java
@@ -1,11 +1,0 @@
-package com.noplanb.item.service;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-@Service
-@RequiredArgsConstructor
-@Transactional(readOnly = true)
-public class ItemService {
-}
