--- src/main/java/com/noplanb/quest/domain/Quest.java
@@ -1,29 +1,0 @@
-package com.noplanb.quest.domain;
-
-import com.noplanb.character.domain.Character;
-import com.noplanb.common.BaseEntity;
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import static jakarta.persistence.GenerationType.IDENTITY;
-
-@Entity
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Quest extends BaseEntity {
-    @Id
-    @GeneratedValue(strategy = IDENTITY)
-    private Long id;
-    private String contents;
-    private Long exp;
-    private Boolean isComplete;
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "character_id")
-    private Character character;
-
-    public void updateCharacter(Character character) {
-        this.character = character;
-    }
-}
