--- src/main/java/com/noplanb/quest/repository/QuestRepository.java
@@ -1,9 +1,0 @@
-package com.noplanb.quest.repository;
-
-import com.noplanb.quest.domain.Quest;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface QuestRepository extends JpaRepository<Quest,Long> {
-}
