--- src/main/java/com/noplanb/user/controller/UserController.java
@@ -1,11 +1,0 @@
-package com.noplanb.user.controller;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-@Controller
-@RequiredArgsConstructor
-@RequestMapping("/api/v1")
-public class UserController {
-}
