--- src/main/java/com/noplanb/user/repository/UserRepository.java
@@ -1,9 +1,0 @@
-package com.noplanb.user.repository;
-
-import com.noplanb.user.domain.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface UserRepository extends JpaRepository<User,Long> {
-}
