--- src/main/java/com/noplanb/user/service/UserService.java
@@ -1,11 +1,0 @@
-package com.noplanb.user.service;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-@Service
-@RequiredArgsConstructor
-@Transactional(readOnly = true)
-public class UserService {
-}
